+++ help/help/src/components/__tests__/unit/Help.tsx
@@ -68,7 +68,7 @@ const mockOnBackButtonClick = jest.fn();
 
 const defaultContentText = 'Default content';
 
-describe('BackButton', () => {
+describe('Help', () => {
   // jest.useFakeTimers();
 
   beforeEach(() => {