+++ editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-mobile-bridge
 
+## 27.0.4
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 27.0.3
 
 ### Patch Changes